refactor(`drasil-srs`): rename primary namespace to `Drasil.SRS`
