+++ code/drasil-example/dblpend/lib/Drasil/DblPend/Requirements.hs
@@ -4,8 +4,8 @@ module Drasil.DblPend.Requirements (
 
 import Language.Drasil
 import Language.Drasil.Document
-import Drasil.DocLang.SRS (datCon)
-import Drasil.DocLang (mkPortableNFR, mkCorrectNFR, inReqWTab)
+import Drasil.SRS.SRS (datCon)
+import Drasil.SRS (mkPortableNFR, mkCorrectNFR, inReqWTab)
 import qualified Language.Drasil.Sentence.Combinators as S
 
 import Data.Drasil.Concepts.Computation (inValue)
