+++ code/drasil-example/gamephysics/lib/Drasil/GamePhysics/Requirements.hs
@@ -7,8 +7,8 @@ import Language.Drasil.Chunk.Concept.NamedCombinators
 import qualified Language.Drasil.Development as D
 import qualified Language.Drasil.Sentence.Combinators as S
 
-import qualified Drasil.DocLang.SRS as SRS (solCharSpec)
-import Drasil.DocLang (mkMaintainableNFR)
+import qualified Drasil.SRS.SRS as SRS (solCharSpec)
+import Drasil.SRS (mkMaintainableNFR)
 import Data.Drasil.Concepts.Documentation as Doc (body, funcReqDom, input_,
   nonFuncReqDom, output_, physicalConstraint, physicalSim, property, solutionCharSpec)
