+++ code/drasil-example/pdcontroller/lib/Drasil/PDController/Body.hs
@@ -3,9 +3,9 @@ module Drasil.PDController.Body (si, mkSRS, pidODEInfo) where
 import Drasil.Database (ChunkDB)
 import Language.Drasil
 import Language.Drasil.Document
-import Drasil.DocLang
+import Drasil.SRS
 import Drasil.Generator (withCommonKnowledge)
-import qualified Drasil.DocLang.SRS as SRS (inModel)
+import qualified Drasil.SRS.SRS as SRS (inModel)
 import qualified Language.Drasil.Sentence.Combinators as S
 import Drasil.System (SmithEtAlSRS, mkSmithEtAlICO)
