+++ code/drasil-example/projectile/lib/Drasil/Projectile/Assumptions.hs
@@ -10,7 +10,7 @@ import qualified Language.Drasil.NaturalLanguage.English.NounPhrase.Combinators 
 import qualified Language.Drasil.Development as D
 import qualified Language.Drasil.Sentence.Combinators as S
 
-import qualified Drasil.DocLang.SRS as SRS (valsOfAuxCons)
+import qualified Drasil.SRS.SRS as SRS (valsOfAuxCons)
 
 import Data.Drasil.Concepts.Documentation (assumpDom, value, consVals)
 import Data.Drasil.Concepts.Math (cartesian, xAxis, xDir, yAxis, yDir, direction, positive)
