+++ code/drasil-example/projectile/lib/Drasil/Projectile/Requirements.hs
@@ -4,9 +4,9 @@ module Drasil.Projectile.Requirements (
 
 import Language.Drasil
 import Language.Drasil.Document
-import Drasil.DocLang.SRS (datCon)
+import Drasil.SRS.SRS (datCon)
 import qualified Language.Drasil.Sentence.Combinators as S
-import Drasil.DocLang (mkMaintainableNFR, mkPortableNFR, mkCorrectNFR,
+import Drasil.SRS (mkMaintainableNFR, mkPortableNFR, mkCorrectNFR,
   mkVerifiableNFR, mkUnderstandableNFR, mkReusableNFR, inReqWTab)
 
 import Data.Drasil.Concepts.Computation (inValue)
