+++ code/drasil-example/sglpend/lib/Drasil/SglPend/Body.hs
@@ -8,9 +8,9 @@ import Language.Drasil hiding (organization)
 import Language.Drasil.Document
 import qualified Language.Drasil.Development as D
 import Theory.Drasil (TheoryModel, output)
-import Drasil.DocLang
+import Drasil.SRS
 import Drasil.Generator (withCommonKnowledge)
-import qualified Drasil.DocLang.SRS as SRS
+import qualified Drasil.SRS.SRS as SRS
 import Language.Drasil.Chunk.Concept.NamedCombinators (the)
 import qualified Language.Drasil.Sentence.Combinators as S
 import Drasil.System (SmithEtAlSRS, mkSmithEtAlICO)
