+++ code/drasil-example/sglpend/lib/Drasil/SglPend/Requirements.hs
@@ -5,7 +5,7 @@ module Drasil.SglPend.Requirements (
 import Language.Drasil
 import Language.Drasil.Document
 import qualified Language.Drasil.Sentence.Combinators as S
-import Drasil.DocLang (inReqWTab)
+import Drasil.SRS (inReqWTab)
 import Data.Drasil.Concepts.Documentation (funcReqDom, output_, value)
 import Drasil.SglPend.IMods (angularDisplacementIM)
 import Drasil.SglPend.Unitals (lenRod, pendDisplacementAngle, inputs)
