+++ code/drasil-example/ssp/lib/Drasil/SSP/IMods.hs
@@ -12,7 +12,7 @@ import Utils.Drasil (weave)
 import Language.Drasil.Chunk.Concept.NamedCombinators
 import qualified Language.Drasil.Development as D
 import qualified Language.Drasil.Sentence.Combinators as S
-import Drasil.DocLang.SRS (propCorSol)
+import Drasil.SRS.SRS (propCorSol)
 
 -- Needed for derivations
 import Data.Drasil.Concepts.Documentation (analysis, assumption, constraint,
