+++ code/drasil-example/ssp/lib/Drasil/SSP/Requirements.hs
@@ -8,9 +8,9 @@ import Language.Drasil.Chunk.Concept.NamedCombinators
 import qualified Language.Drasil.Development as D
 import qualified Language.Drasil.Sentence.Combinators as S
 
-import Drasil.DocLang (mkInputPropsTable, mkMaintainableNFR, mkCorrectNFR,
+import Drasil.SRS (mkInputPropsTable, mkMaintainableNFR, mkCorrectNFR,
   mkUnderstandableNFR, mkReusableNFR)
-import Drasil.DocLang.SRS (datCon, propCorSol)
+import Drasil.SRS.SRS (datCon, propCorSol)
 
 import Utils.Drasil (mkTable)
