+++ code/drasil-gen/lib/Drasil/Generator/CaseStudyVariants.hs
@@ -19,7 +19,7 @@ import GHC.IO.Encoding (setLocaleEncoding, utf8)
 import Drasil.FileHandling (FileLayout, OverwritePolicy(..), directory, localPath, ps,
   writeFiles)
 import Drasil.LessonPlan (LsnDesc)
-import Drasil.DocLang (SRSDecl, mkDoc)
+import Drasil.SRS (SRSDecl, mkDoc)
 import Drasil.System (DrasilWebsite, LessonPlan, SmithEtAlSRS, programName)
 import Language.Drasil.Document (Document)
 import Language.Drasil.Code (Choices)
