+++ code/drasil-srs/lib/Drasil/SRS/SRS.hs
@@ -4,7 +4,7 @@
 -- in Drasil' tutorial found on the wiki:
 -- https://github.com/JacquesCarette/Drasil/wiki/Creating-Your-Project-in-Drasil
 
-module Drasil.DocLang.SRS (
+module Drasil.SRS.SRS (
   -- * Section Constructors
   -- | For use in an SRS document. Ordered by appearance in a SRS.
   tOfCont, refMat, tOfUnit, tOfSymb, tOfAbbAcc, intro, prpsOfDoc, scpOfReq,
