+++ code/drasil-srs/lib/Drasil/Sections/AuxiliaryConstants.hs
@@ -12,7 +12,7 @@ import Language.Drasil.Document
 import Utils.Drasil (mkTable)
 
 -- Other docLang
-import qualified Drasil.DocLang.SRS as SRS (valsOfAuxCons)
+import qualified Drasil.SRS.SRS as SRS (valsOfAuxCons)
 import Drasil.DocumentLanguage.Units (toSentence) -- TODO: suspicious
 import Drasil.Sections.ReferenceMaterial (emptySectSentPlu)
