+++ code/drasil-srs/lib/Drasil/Sections/SpecificSystemDescription.hs
@@ -49,7 +49,7 @@ import Drasil.Metadata.TheoryConcepts (inModel, thModel, dataDefn, genDefn)
 
 -- other docLang
 import Drasil.DocumentLanguage.Definitions (helperRefs)
-import qualified Drasil.DocLang.SRS as SRS
+import qualified Drasil.SRS.SRS as SRS
 import Drasil.Sections.ReferenceMaterial(emptySectSentPlu)
 
 -- Takes the system and subsections.
